Change create_fmod_guid_map FMOD path check
Now checks if the length of the path string is greater than 7 instead of if the string is empty, since it should never be empty. "event:/" is not a valid FMOD path since empty event names are prohibited. Technically FMOD has additional restrictions on event names such as prohibiting the characters <>:"/\|?*, but it shouldn't matter here.